+++ app/src/main/java/com/rxjava2/android/samples/ui/OperatorsActivity.java
@@ -8,19 +8,19 @@ import android.view.View;
 import com.rxjava2.android.samples.R;
 import com.rxjava2.android.samples.ui.operators.AsyncSubjectExampleActivity;
 import com.rxjava2.android.samples.ui.operators.BehaviorSubjectExampleActivity;
-import com.rxjava2.android.samples.ui.operators.BufferExampleActivity;
+import com.rxjava2.android.samples.ui.operators.transform.BufferExampleActivity;
 import com.rxjava2.android.samples.ui.operators.CompletableObserverExampleActivity;
-import com.rxjava2.android.samples.ui.operators.ConcatExampleActivity;
-import com.rxjava2.android.samples.ui.operators.DebounceExampleActivity;
-import com.rxjava2.android.samples.ui.operators.DeferExampleActivity;
-import com.rxjava2.android.samples.ui.operators.DelayExampleActivity;
+import com.rxjava2.android.samples.ui.operators.mathematical.ConcatExampleActivity;
+import com.rxjava2.android.samples.ui.operators.filter.DebounceExampleActivity;
+import com.rxjava2.android.samples.ui.operators.create.DeferExampleActivity;
+import com.rxjava2.android.samples.ui.operators.utility.DelayExampleActivity;
 import com.rxjava2.android.samples.ui.operators.DisposableExampleActivity;
-import com.rxjava2.android.samples.ui.operators.DistinctExampleActivity;
-import com.rxjava2.android.samples.ui.operators.FilterExampleActivity;
+import com.rxjava2.android.samples.ui.operators.filter.DistinctExampleActivity;
+import com.rxjava2.android.samples.ui.operators.filter.FilterExampleActivity;
 import com.rxjava2.android.samples.ui.operators.FlowableExampleActivity;
-import com.rxjava2.android.samples.ui.operators.IntervalExampleActivity;
-import com.rxjava2.android.samples.ui.operators.LastOperatorExampleActivity;
-import com.rxjava2.android.samples.ui.operators.MapExampleActivity;
+import com.rxjava2.android.samples.ui.operators.create.IntervalExampleActivity;
+import com.rxjava2.android.samples.ui.operators.filter.LastOperatorExampleActivity;
+import com.rxjava2.android.samples.ui.operators.transform.MapExampleActivity;
 import com.rxjava2.android.samples.ui.operators.MergeExampleActivity;
 import com.rxjava2.android.samples.ui.operators.MeterailizeExampleActivity;
 import com.rxjava2.android.samples.ui.operators.PublishSubjectExampleActivity;
