+++ app/src/main/java/com/rxjava2/android/samples/ui/operators/utility/DelayExampleActivity.java
@@ -1,4 +1,4 @@
-package com.rxjava2.android.samples.ui.operators;
+package com.rxjava2.android.samples.ui.operators.utility;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -13,20 +13,20 @@ import com.rxjava2.android.samples.utils.AppConstant;
 import java.util.concurrent.TimeUnit;
 
 import io.reactivex.Observable;
+import io.reactivex.Observer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.observers.DisposableObserver;
+import io.reactivex.disposables.Disposable;
 import io.reactivex.schedulers.Schedulers;
 
 /**
- * Created by amitshekhar on 27/08/16.
+ * Created by amitshekhar on 05/03/17.
  */
-public class IntervalExampleActivity extends AppCompatActivity {
 
-    private static final String TAG = IntervalExampleActivity.class.getSimpleName();
+public class DelayExampleActivity extends AppCompatActivity {
+
+    private static final String TAG = DelayExampleActivity.class.getSimpleName();
     Button btn;
     TextView textView;
-    private final CompositeDisposable disposables = new CompositeDisposable();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -43,34 +43,32 @@ public class IntervalExampleActivity extends AppCompatActivity {
         });
     }
 
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        disposables.clear(); // clearing it : do not emit after destroy
-    }
-
     /*
-     * simple example using interval to run task at an interval of 2 sec
-     * which start immediately
+     * simple example using delay to emit after 2 second
      */
     private void doSomeWork() {
-        disposables.add(getObservable()
+        getObservable().delay(2, TimeUnit.SECONDS)
                 // Run on a background thread
                 .subscribeOn(Schedulers.io())
                 // Be notified on the main thread
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribeWith(getObserver()));
+                .subscribe(getObserver());
     }
 
-    private Observable<? extends Long> getObservable() {
-        return Observable.interval(0, 2, TimeUnit.SECONDS);
+    private Observable<String> getObservable() {
+        return Observable.just("Amit");
     }
 
-    private DisposableObserver<Long> getObserver() {
-        return new DisposableObserver<Long>() {
+    private Observer<String> getObserver() {
+        return new Observer<String>() {
 
             @Override
-            public void onNext(Long value) {
+            public void onSubscribe(Disposable d) {
+                Log.d(TAG, " onSubscribe : " + d.isDisposed());
+            }
+
+            @Override
+            public void onNext(String value) {
                 textView.append(" onNext : value : " + value);
                 textView.append(AppConstant.LINE_SEPARATOR);
                 Log.d(TAG, " onNext : value : " + value);
